Fix bug where GraphPoints2D function would stop half way.
Whenever two identical points existed, the output function would
skip the rest of the points.

Also, we were outputting blank spaces at the end of the line,
in GraphPoints2D, due to a < instead of a >.

Finally, made the GraphPoints2D function compile by default,
since they're useful, and provide an option to disable their
compilation in the cmake build script.

Tested: Graphs print properly when running tests on Linux and Mac.
Change-Id: I4546ad4b03fbef282cfce21f30273b4dd70180f2