Fix crash on anim load.
We held a pointer in MatrixAnim's SplinePlayback to a
structure that was created on the stack. Fixed by ensuring
the structure is a member variable in MatrixAnim.

Also, the vector of structures would be realloc'd and
the internal pointers would become invalid. Have to
avoid realloc'ing this vector by pre-calculating its length.

These are temporary fixes. MatrixAnim will be reimplemented
in a later change to be rebust and more efficient.

Tested: Works in test game.
Change-Id: I292a67ace5abdace5c72543f0eb391e85f93e2d5