Improve precision of QuadraticCurve::Roots().
This function wasn't properly handling the case where all
coefficients were very small (< 2^-7) or very large (> 2^7).

This case is problematic because we multiply by the reciprocal
of a coefficient during the root calculation. That means we'll
multiply a large number by a small number and the precision
will disappear.

The fix is to detect this case and normalize the coefficients
so that the largest is near one.

Note that we already properly handle the case where one or more
of the coefficients is tiny compared to the largest.

Tested: Added a new test case for really small values.
Change-Id: I403ffbab6230ee0d0bd69b3152ea286cba9470a4