Remove translate from skeleton nodes.
The translate was also being applied in the mesh_pipeline,
because "Freeze Transforms" doesn't clear the translate
on joints. The result is that the translate was applied
twice: once in the constant transform output by the
mesh_pipeline, and once in the constant translate output
in the animation data.

Tested: Works on animation in sample app.
Change-Id: Ia81e9f61e1bab601b6d020d0eafa1bce9db95898