Add 'repeat' flag for animations that can cycle.
The anim_pipeline detects when an animation start and ends in the
same state. For these animations it sets the `repeat` flag to true
in the fplanim file. On load, when `repeat` is true, the animation
splines are initialized to repeat, so that the animation runs
forever.

Change-Id: I87ca5601519a1fecbd7743fecbbf982f75cb0eca
Tested: Works on test animations.