Blend rigged animations.
The SplinePlayback now takes a blend_x parameter to
specify the blend time. The blend time is used by
the BulkSplineProcessor to smoothly transition from
the current state to the new spline.

In a future change we will pull the SplinePlayback
out of the MatrixAnim class. This will let you
configure the playback parameters per call.

Tested: Works with sample game on Linux and Android.
Change-Id: Iaa80e48a983dc2d40eeb457d54be4f16c948d7e1