Fix bug in anim_pipeline that could mark channels as const.
We weren't checking the last node of a spline, to see if it differed
in value. So, if all the nodes were the same except for the last,
and all the derivatives were 0 (including the last), the channel
would be erroneously marked as `const`.

Further, if the `const` value was the default value, the channel
would be deleted.

This bug resulted in some scale channels being deleted from some
animations. The channel started at scale 1, and ended at scale 0
(making the character disappear), but the entire channel was
deleted, so the character scale would not actually change in the
output animation.

Change-Id: Idd0deb1b125f56516c6f40ca006e8fd3db39d85f
Tested: Works on animations described above, on Mac and Linux.