Improve animation curve comparison check.
The derivative check was previously a function of the
total length of the curve, which leads to unexpected results.

We replace that with an absolute comparison of the
derivatives, converted to angles.

This seems to work a lot better on our data.

Change-Id: I5d11008d9270b8e067486dc14da7963d3882351d
Tested: Works on Mac and Linux on sample game's animations.