Add --repeat and --norepeat options to anim_pipeline.
Allows the caller to explicitly mark an animation as
repeating or non-repeating. Some animations start and end
in the same state, but should not be repeated. Others
should repeat even if they have a slight glitch.

If there will be a glitch during the repeat, a warning
is issued.

Change-Id: I535691f180924fb1a0447bad9eedd884ba0f00ba
Tested: Works on sample app's animations, on Mac and Linux.