Support animations with 360 degree rotations.
Previously, we normalized all angles to be between -180 and +180
degrees. This doesn't work when an aniamtion does a flip, all the
way around. For flip animations, the rotation goes from, say,
0-->360.

The solution is **not** to normalize the angles. We also don't
normalize the output of the BulkSplineEvaluator. The client of
the output is now responsible for normalizing.

This is a nice optimization, because many times the client doesn't
care if an angle is normalized or not. The client will just take
sin or cos of the angle, and if it's somewhat outside the +-180
degree region, it doesn't affect the math significantly.

It also significantly simplifies BulkSplineEvaluator::EvaluateCubics,
which is nice because that function has assembly language
implementations.

Change-Id: If3d6ba411203629298bb1f120774a48599ba8000
Tested: Works on Mac and Linux.