Add functions to log a rig's animation in CSV format.
Comma-separated-values are useful because they can be
loaded into a spreadsheet and easily visualized.

RigMotivators now expose a csv-header function that
outputs a string to label all the bones and operations,
and a csv-values function that outputs the rig's
component values, at the current time.

Super useful for debugging. Shouldn't be called when
not debugging, since slow.

Added some documentation, too.

Tested: Works well on Linux and Android.
Change-Id: Ib94f5cc31daed8e193ff6c33fd5f21262044281c