Fix bug where the spline y-range could be too small.
For modular splines, we need to expand the spline's
y-range to match the number of nodes in the spline.
It's possible for the spline to keep going up and up.

Note that we only normalize the first value of the spline,
and subsequent ones are allowed to curve out of the
normalized range.

Change-Id: I5b7a84646f217f2203fa7e47fce3f86d3730a126
Tested: Assert no longer fires, on Linux.