Fix anim_pipeline bug that caused channels to be const.
If an animation channel had two points that shared the
same derivative, and the first point didn't start at
time zero, and the channel was marked as cyclic, then
the channel would be considered constant.

The fix is to look at the first key's value instead of
the value at time 0 (which can be different).

Change-Id: I12478808682d13e020f13679528c1b2367c35042
Tested: Fixes bug with sample app animation