Propagating uniforms mat4x3 to the RigProcessor.
In order to save uniforms, the `bone_transforms` now use 3x vec4's in
the shader instead of the traditional mat4. In C++, we now represent
this using a `mathfu::Matrix<float, 4, 3>` or `typedef AffineTransform`.

This change needed to be reflected in the `RigProcessor`, so all
affected `motive` files have been updated accordingly.

Bug: b/24951383
Change-Id: I8625f48d4e9c0e4f15970d4552c326b0d686d6d6
Tested: Tested by running Zooshi on Ubuntu.