Pull spline out of SplinePlayback.
Allows us to specify animation start time, playback
rate, and blend time when launching an animation.

Previously (for legacy reasons), this information was
stored per-spline, which is a silly waste of memory.

Now we don't hold a copy of SplinePlayback anywhere.
It's passed in when the animation is launched.

In addition to being a good API clean-up, this is
required to initialize to an animations that differs
from its default pose. A blend time of 0 must be
specified for those animations.

Tested: Works well on OSX.
Change-Id: Ic25c05a65d1020ed8c33ba1e29e89aa9c2793a9e