Fix unit tests.
The BulkSplineEvaluator no longer normalizes values before
returning them. In some applications it's preferable that
we don't. For example, we'd have to insert nodes and
discontinuities at y=+-pi when generating anim_pipeline data.
So we just delete this test from spline_test.

The spline pointers have been pulled out of SplinePlayback.
This changes the API a bit and caused a couple build errors.

By coincidence, we were getting exact values when quantizing
test values before, so our precision was very high. Nothing
has changed internally, precision wise, but the quantize range
is now different. This means that the lower precision (which
always existed) is now exposed in the tests. We lower the
precision check in the test to compensate.

Tested: Tests now build and pass on Linux.
Change-Id: I8a96306c76628ab3da16736623b3161833d58a1d