Separate sanity check from validity check.
Calling Motivator::Valid() is common and should happen
as quickly as possible.

Motivator::Sane() is an internal consistency check and
should never fail. Having this check inside Valid() was
redundant, but nice during initial development, where
it caught some tricky bugs early. We haven't had an internal
consistency bug in almost a year, though, and we can always
add it back in by calling Sane() all over the code, if we
need to.

Change-Id: If9495610a8b2d1c4522271063fcc714b7ea1ce0f
Tested: Unit tests build and run fine on Mac.