Fixing android build for Motive.
There was a rename from `smooth1f` to `spline1f`, but the Android.mk
related files were never updated, so the build was failing.

The docs for building on android were also updated.

Bug: b/26113750

Tested: Tested by building for android on OSX and Linux. Also
tested the docs using generate_docs.py and viewed on Chrome.

Change-Id: I54e556ab32a60c8b995a5e6b93fee1e1e3873f60