Reference flatbuffers lib instead of compiling sources.
We used to include the flatbuffer .cpp files directly, but now there
is a `flatbuffer` target in both Android and cmake. Much better to
depend on that since if a new file .cpp file is added to FlatBuffers
the dependent builds won't break.

Change-Id: I8c345758a82e39a537f5dec69752d00e84789796
Tested: unit tests build and run on Linux.