Add fplviewer tool to preview fplmesh and motiveanim files.
The viewer loads an fplmesh file and lets the user move the
camera around it with the mouse.

It also lets the user play an animation back on that mesh,
and dumps all the debug information for that animation to
a csv file.

The mesh can have textures or be untextured.
The mesh can be skinned or not skinned.
The animation can animate only the root bone, or it can
animate a full rig.

Expand the Angle class with a functions that return the
up vector for a coordinate system, and the unit sphere
point for a geographic coordinate.

Tested: Works on Linux and Mac. Android still compiles.
At the moment, not intended for mobile.

Change-Id: I2ef3d6f9a78ed76599876cf174c1f3750c91947c