Fix anim_pipeline bug when some channels ended early.
If a channel ended early but its end derivative was not zero,
then it would be extended to the animation end time, but would
curve to that end time. We want instead for it to hold its
last value until it reaches the end.

This manifested itself in an animation where scale went to zero.
When the scale curve was extended, it dipped below zero and became
negative.

The fix is to add a point at the existing end of the curve with
derivative 0, in addition to the point at the end of the animation.

We need to update CompactSpline to accept two points at the same x
location. This plays back as valid. It's just a discontinuity (in
this case, in the derivative).

Bug: 29208446
Tested: Fixes animation described above, on Linux and Mac.
Change-Id: I446ea315ed1dc7a6d7fa177d4b683e727e5d8691