anim_pipeline now has options to convert unit and axes.
These options were already in mesh_pipeline, and are needed in
anim_pipeline in order for the output of the two pipelines to match.

Don't submit new binaries just yet because the next change
(to put vertices in object space) is required to make
this change effective.

Change-Id: I6e98a0cbb767fe26ae33b2ec2a3190c216981ff4
Tested: Builds and runs on Linux.