Allow target time to be 0 in Target1f().
A target time of 0 means change the current values. We asserted on this condition since the Current1f() function is used to set the current values. However, there's no reason why Target1f() can't be used to set the current values, and sometimes that degenerate case is just what we want.

From cl/127105950.

Change-Id: I23d17dcfa9b540c4358973f422e7bd3c433c27a7