Add --root_bones_only option to anim_pipeline.
This option outputs only the top-most bone with animation data.
Each such bone is output to its own file with the base name +
bone name, unless there is only one such bone, in which case
the output file name is the standard name.

Change-Id: Ie6d501b46f6ccb1f03f1569aaf52062b18d91715
Tested: Works for test-file with two root-bone animations, on Linux.