Fix bug where animation would be stuck at end.
If delta_time was ever greater than the animation's length,
the animation would stay stuck at the end. This was caused
by the simple subtraction instead of the mod.

Replaced with a Range function that does the subtraction,
checks it, and if it still fails, does the mod. It's a handy
function because it is still fast in the most common case
where the subtraction just works.

Change-Id: I80e413210b365d5e601e379b4130f27f226f71cf
Tested: Fixes lullaby animation stuckage.