Allow AnimTable to load from embedded RigAnims.
Maintain legacy interface so that existing data is not broken.

Allows us to create an AnimTable file that has all of its
animations inside of it. This is useful for two reasons,
- keeps file count down, for operations like --root_bones_only
  which output a different RigAnim for each object in the scene
- allows pipelines to know exactly what the output will be
  from anim_pipeline, since now we can have a one-output for
  one-input situation in --root_bones_only

Also simplified and improved the LoadFn signature so that
we don't have to copy the data into scratch_buf if it's already
been loaded.

Tested: Added new unit tests for this functionality.
Change-Id: Ief778038098cb8cb37320ee30fc1a69e50be7291