anim_pipeline now starts animations at time 0, by default.
Previously, animations often started at frame 1 instead of frame 0.
This is the default in Maya, so it's very common.
This resulted in a 33 tick gap at the start of many animations.
By default, the animation will just be shifted over to eliminate the
gap.

Also, it's possible for animation curves to be authored with negative
times. This is supported by FBX, and some animations are authored
that way (especially cyclical animations). These animations were
causing anim_pipeline to assert. We now shift them, as well, to start
at time 0 so that there are no negative times exported.

We also add an option to disable the shifting --preserve_start_time.
If negative times exist and this option is enabled, we simply clamp
the negative times to 0 and output a warning.

Tested: Works with sample file that has several negative channels
starting at different times. Zooshi also runs fine, on Linux.

Change-Id: I2b58f56cecba606865e00cb5d9020ae0431c0734