Add CompactSpline::InitFromSpline() and InitFromNodes().
These functions allow a CompactSpline to be initialized from
another CompactSpline (that's evaluated at constant x intervals),
or from an array of uncompressed nodes.

This is usefull if cubic math is too complicated and you prefer
to brute force a solution over many points. In that case, you
can create a new CompactSpline that is a fine grained evaluation
of the original, and do the math point-wise on it.

Tested: New unit tests pass on Mac and Linux.
Change-Id: I404f83e33df5f8b313038e475e560fc575141bb9