Fix bug in derivative evaluation at end of spline.
When the x was right as the end of the spline, the derivative would
be 0 instead of the derivative of the last node. This was because
we were initializing the spline with the "beyond end" state, in which
the derivative is always 0. However, when right at the end we should
initialize for the last segment so that we get the derivative for
the last node.

The fix is to special-case the index lookup when x is right at the
end of the node.

Change-Id: I9d2d451ccf9de1a8266650446897f7fd887e81bd
Tested: Updated unit tests to be correct, on Mac.