Add math for ease-in ease-out curve calculation.
Calculates the x intersection point of two quadratic curves, f(x) and
g(x), of opposing curvature and overlapping range and the shift
needed to place g(x) on the right of f(x).

Does the calculation by fixing an intersection point u and finding
shift s by setting the discriminant of h(u) = f(u) - g(u) to 0.
We solve for h(u) = 0 and take the larger root as our shift. We then
use the quadratic formula with a 0 discriminant to solve for u, our
intersection point.

Supports regular cases where the intersection is between the start
and total x, as well as special cases where the intersection is
either before start x or after total x. In special cases, only one
curve is used and the start value, start derivative, and end value
are achieved, while we get as close as possible to the end
derivative. In these cases, in_x and total_x are equal and
calculated by finding the x value at which the curve hits the end
value.

This math will let us interrupt a curve part-way through to go to a
new target, which is important for responsive animations. It is also
stable, so that if we interrupt every frame to the same target, it
will maintain the original curve. Creating a stable curve is
difficult, but we've added extensive unit tests for it.

Tested: New unit tests pass on OSX, Linux, and Android.

Change-Id: Ia2afc100a940f87d622d712071a4e46dce7fc109