Factor out floating point utility functions.
These functions are useful when dealing with numerical precision.
Power of two multiplication maintains the mantissa, so there is
no loss of precision.

Note that these functions only work if float is the IEEE 32-bit
floating point type. This is true for all compilers that we target.
A static_assert() exists to double-check this at compile time.

Change-Id: I4ddfbd1fcd2aa9270dd994d30e2154303274daeb
Tested: New unit tests pass on OSX, Linux, and Android.