Improve precision of QuadraticCurve::Roots().
Scales the quadratic curve in both the x and y axes before applying
the quadratic formula so that,
- (from scale in x) the x^2 coefficient is near the larger of the x
  and constant coefficients,
- (from scale in y) the x^2 coefficient is near 1.
By keeping the coefficients near 1, especially the x^2 coefficient
(which is in the denominator of the quadratic formula), precision is
increased.

Geometrically, you can imagine this as putting the quadratic curve
in a nice round shape--not too flat and not too skinny. When the
shape is balanced, all numbers are in the same scale so precision is
easy to maintain.

Note that care has to be taken when numbers are near max float and
min float, because it's possible to scale to infinity.

Tested: Created new unit tests. Upgraded some unit tests to expect
two roots instead of 1, because the extra precision can now find
both results. New tests pass on Mac, Linux, and Android.

Change-Id: Ic8cc4dc9791d95fdf38b1e47ab2bf65e9a6ebeef