Fix crash when spline is nullptr.
If we set an existing spline to be null, and then that curve
gets reinitialized, it will crash when it reaches the end of
the cubic and needs to be reinitialized. Here's the callstack:

CompactSpline::NodeX() <== here CompactSpline is null
BulkSplineEvaluator::CubicStartX()
BulkSplineEvaluator::X()
BulkSplineEvaluator::AdvanceFrame()

The fix is to safely reset splines to null, and to ensure that
we never try to reinitialize a cubic for a spline that's null.

Change-Id: Ie45256d207375a61c7c3e70d7d85a0970f59a0f3
Tested: Zooshi runs fine on Linux.