Fix crash from reading uninitialized nodes past the end of a spline.
We were returning the last node when the time matched perfectly,
even if there was only one node in the spline. This caused all
sorts of errors because we would read past the end of the segments
array to initialize the segment.

Change-Id: I760b0e44c866e8af44c8d1cbc3b053d6265d0274
Tested: Fixes motive_test unit tests, on Linux.