Add processor for ease-in ease-out animation.
Adds an EaseInEaseOutProcessor for the ease-in ease-out curve math.
This math allows for stable and responsive animation by allowing
us to interrupt a curve part-way through to go towards a new
target.

More details about the math can be found in the change
titled "Add math for ease-in ease-out curve calculation.".

Tested: New unit tests work on Android, Linux, and OSX.

Change-Id: Ib35811f6cadd3d5a6fabb1aeec5c93655e3db102