Add interface to ease-in ease-out curve calculation.
Calculates a start and end second derivative using the input typical
distance, typical time, and bias. The curve is then constructed
using the calculated start and end second derivative, user-input
start and end derivative, and user-input start and end value.

The second derivatives are calculated using zero-value start and
end derivatives, typical total x (u), typical delta value (d), and
bias (S).

We let f(x) and g(x) represent our ease-in and ease-out
curves with second derivatives 2A and 2a respectively. Since we
assume zero-value derivatives, we know that f(0) = 0 and f'(0) = 0,
which gives us: f(x) = Ax^2.

For a bias of 0.5f, we have a balanced curve and by symmetry, travel
half the distance in half the time. For a bias of 1.0f, we travel
the full distance in full time and disregard the end second
derivative.

By scaling A linearly between the two formulas obtained for the bias
at 0.5f (A = 2d / u^2) and 1.0f (A= d / u^2), we can get the
equation A = d / (SU^2) for 0.5f <= S <= 1.0f.

We then use g(x) = ax^2 + bx + c and the knowledge that g(u) = d
and g'(u) = 0 to rewrite the equation for g in terms of a.

We use our equation for f(x) written in terms of A with our equation
for g(x) written in terms of a to solve for a by adjusting the
curves so that g and f intersect once and setting the discriminant
of the adjusted equation to 0.

Once we solve for A and a, we double those values to obtain
our calculated start and end second derivatives. By symmetry,
if our bias is less than 0.5f, we can use 1 - S for our calculations
and flip our end results.

This math will provide users with a simpler interface for
generating curves, as second derivatives are not intuitive for
people to imagine. By leaving the second derivatives to the math,
users will be able to create a curve with just by specifying their
desired start/end values/derivatives, approximately where they
want the in and out curves to intersect, and the typical distance
and time it should take.

Tested: New unit tests pass on OSX, Linux, and Android.

Change-Id: Iac30342faae6ab489f5541e2af87f1cacf28fca0