Various refactors to the motive library.
Remove MatrixOpArray class which is a simple wrapper around a
vector<MatrixOperationInit>.  Instead, just using the vector
directly everywhere simplifies a lot of the code (and removes
one layer of indirection).

Make CalculateResultMatrix a static function of the MatrixOperation
class instead of a member function of the MatrixData class.

And finally branch the index_allocator from fplutil to reduce motive
dependencies.

PiperOrigin-RevId: 184695570
Change-Id: Ia5678ca24f3e6e837e33371da6103de1fa90fa56