Roll-forward.
Resubmitting original change with a default initialization step.

MatrixOps::BlendToOp() checks the Value() and Velocity() of the current spline when blending in an effort to collapse splines into constants. Velocity() fails if the spline is uninitialized.

*** Original Change ***

Remove the need for defining animations.

This required some fundamental changes to low-level motive constructs. Most of the changes remove code included to support older C++ compilers, in particular old versions of Visual Studio (VS2010~2012).

Motivator:
- Use the move ctor and assignment ops correctly. VS2010~2012 required move...

***

PiperOrigin-RevId: 198632612
Change-Id: Ic774995994a312fafefe3f5829f13f21f9c9a396