Sanity check constant values in the anim_pipeline.
This function returns true if the given curve is actually a constant value, meaning every node in it has the same value and they all have 0 derivatives.

It appears possible that FbxAnimCurveNode::GetChannelValue() can return the wrong value from some binary FBX files. The value of every point on the curve should be the same as the value from GetChannelValue(), so a sanity check at the end saves this case.

PiperOrigin-RevId: 201355665
Change-Id: I8d792fe11465336296970f021d5b8bd5a6503b82