Give Processors a pointer to the owning MotiveEngine.
This is done via a setter instead of the constructor primarily for backwards compatibility since Processors are constructed by a template-generated function and adding it to the MotiveProcessor constructor would require all Processors to have an additional or updated constructor.

Additionally, SetEngine() is called directly after construction and has no effect after being called once, so the value is effectively set at construction time.

PiperOrigin-RevId: 208065277
Change-Id: I186c07050821107a845fd3d8e85efffa17415726