Add support for extracting root motion from rig animations.
Specifying a "root motion bone" at initialization time does two things:
1. When calculating bone transforms, the root motion bone will have it's transform set to the identity.
2. The transform from step 1) will be available via RigMotivator::RootMotionTransform().

The bone is specified by index because bone names are not always saved. If no bone is specified or the bone is invalid, RootMotionTransform() always returns identity and no bone transforms are modified.

Note that this simply extracts the root motion but does not actually *do* anything with it. If a caller provided a valid root motion bone at initialization time but then did not use RootMotionTransform() somewhere, it would effectively strip root motion from the animation.

PiperOrigin-RevId: 209682438
Change-Id: I1203485974c9e2d920f020f8be03f2f5aeec7c22