Add duplication to the Motivator API and implement it for the SplineProcessor.
Duplication is explicitly not done via the copy constructor and assignment operators since the new Motivator is not a true "copy" of the old one, but rather initialized to the current state of the old one.

The actual logic is in the new InitializeIndices() function that Processors are responsible for implementing. Processors must "opt in" to duplication functionality by overriding SupportsDuplication().

This CL includes an implementation for the SplineProcessor. I'll do the MatrixProcessor in a follow-up CL.

PiperOrigin-RevId: 211097431
Change-Id: Id3292664c0913657ac6b62aab78dc3bed4cef4ff