Unwind recursion in animation pipelines.
Both pipelines use the AddCurve() function to recursively break down curves into smaller segments when a large segment will have a significant error. This recursion can go pretty deep and cause a stack overflow in some environments, so unwinding it in favor of a loop is beneficial.

This should not change the output animations in any way - it simply replaces recursion with a loop and a data structure for bookkeeping.

Also fixes a potential array-out-of-bounds issue in GenerateDerivatives().

PiperOrigin-RevId: 213702215
Change-Id: I2af9a36fedcc956cfe49a7678834fd790c8a77c3