Add quaternions and the SqtProcessor to Motive.
The SqtProcessor is an alternative implementation of the MatrixProcessor API that is backed by translation, quaternion rotation, and scale components. It can be used anywhere a MatrixMotivator is used but processes MatrixOperations differently. It uses MatrixOperations to ensure RigData can support either Sqt-backed or Matrix-backed animations.

Because of how Motive sets up Motivators, this required a bit of extra work in the RigData struct to allow a RigMotivator to be backed by Matrix-based animations or Sqt-based animations without having to "think about it" (outside of initialization time).

PiperOrigin-RevId: 223064674
Change-Id: Idff3afd91a36f85a2401ee2c908d8656cc764526