Rollback.
*** Original change description ***

Add quaternions and the SqtProcessor to Motive.

The SqtProcessor is an alternative implementation of the MatrixProcessor API that is backed by translation, quaternion rotation, and scale components. It can be used anywhere a MatrixMotivator is used but processes MatrixOperations differently. It uses MatrixOperations to ensure RigData can support either Sqt-backed or Matrix-backed animations.

Because of how Motive sets up Motivators, this required a bit of extra work in the RigData struct to al...

***

PiperOrigin-RevId: 223098676
Change-Id: Ia2df73398fd86f780fdf5d219c558abb14f5a8db