Rollforward.
*** Original change description ***

Add quaternions and the SqtProcessor to Motive.

The SqtProcessor is an alternative implementation of the MatrixProcessor API that is backed by translation, quaternion rotation, and scale components. It can be used anywhere a MatrixMotivator is used but processes MatrixOperations differently. It uses MatrixOperations to ensure RigData can support either Sqt-...

***

PiperOrigin-RevId: 223397272
Change-Id: Id78304faf97b95a8732ac9f86286190e8f834417