Add a flag indicating whether or not a MatrixAnim is an Sqt-based animation.
This is required to properly initialize MatrixMotivators in an Sqt-based rig animation for bones that have no rotations at all (only translation and scale). Previously, we marked a MatrixAnim as Sqt *only* if it had quaternions. This doesn't cover the case of an Sqt-compiled animation with no rotation at all and results in a crash when blending to one that actually has rotation in that bone.

PiperOrigin-RevId: 226043810
Change-Id: Ib69aba11f635a12e3824ea0d2a1d6bfb0d4c48a1