Take into account y_scale and y_offset when calculating blend init.
The blend spline needs to take into account that the spline it is blending to may be scaled and shifted. As soon as the blend spline is complete, the next CubicInit will be using the y_scale and y_offset, so this only needs to be done for the BlendInit.

PiperOrigin-RevId: 229201677
Change-Id: I6747ab71aea03a22d28528a1e5bae91de6dff11d