Filter quaternion flips in keyframe conversion utility.
This handles animation fidgeting that results from the fact that two different quaternions, q and -q, can represent the same orientation.

For example, (1,0,0,0) and (-1,0,0,0) are both technically the same orientation (the identity rotation). However, averaging them on a component-wise basis gives (0,0,0,0), and interpolating with any other value gives a non-identity rotation. We can avoid this problem by ensuring that all neighboring keyframes have quaternions that are in the same 4-dimensional hemisphere, meaning that they are as "close" together as possible. Mathematically, this means ensuring that their dot product is non-negative.

We do something similar in the SqtProcessor's implementation.

PiperOrigin-RevId: 239312762
Change-Id: Ia8d2436a9cef331a041ee2594910bb43a03ebbeb