Fix ubsan failure, divide-by-zero in Range::PercentClamped
For quaternion curves, the components of a quaternion are evaluated separately.
In some cases, it's possible for all y-values of a spline to be the same, such
as for the AnimatedCube gltf sample which has an animation in the following
form with kOrderXYZW:

f(0) = 0, 0, 0, 1
f(1) = 0, 1, 0, -4.37114e-08
f(2) = 0, -8.74228e-08, 0, -1

The quat.x component is always 0, so it results in the y_range of 0->0.  When
converting to CompactY, the y-value is converted to a percentage, which hits a
divide-by-zero with y/y_range.Length().  To prevent this, add a special-case for
a zero-length range.

PiperOrigin-RevId: 246067151
Change-Id: I546d3d4d89e1d5563de6920a864901e20ea9ba6b